requirements.txt
acdh-tei-pyutils==0.17.0
Jinja2==3.0.3
python-slugify==5.0.2